+++ app/containers/Toast.js
@@ -3,22 +3,23 @@ import { StyleSheet } from 'react-native';
 import EasyToast from 'react-native-easy-toast';
 
 import { COLOR_TOAST, COLOR_WHITE } from '../constants/colors';
-import { isNotch } from '../utils/deviceInfo';
 import sharedStyles from '../views/Styles';
 import EventEmitter from '../utils/events';
 
 const styles = StyleSheet.create({
 	toast: {
-		backgroundColor: COLOR_TOAST
+		backgroundColor: COLOR_TOAST,
+		maxWidth: 300,
+		padding: 10
 	},
 	text: {
 		...sharedStyles.textRegular,
 		color: COLOR_WHITE,
-		fontSize: 14
+		fontSize: 14,
+		textAlign: 'center'
 	}
 });
 
-const POSITION_VALUE = isNotch ? 230 : 200;
 export const LISTENER = 'Toast';
 
 export default class Toast extends React.Component {
@@ -26,22 +27,26 @@ export default class Toast extends React.Component {
 		EventEmitter.addEventListener(LISTENER, this.showToast);
 	}
 
+	shouldComponentUpdate() {
+		return false;
+	}
+
 	componentWillUnmount() {
 		EventEmitter.removeListener(LISTENER);
 	}
 
 	showToast = ({ message }) => {
-		this.toast.show(message);
+		this.toast.show(message, 1000);
 	}
 
 	render() {
 		return (
 			<EasyToast
 				ref={toast => this.toast = toast}
-				positionValue={POSITION_VALUE}
+				position='center'
 				style={styles.toast}
 				textStyle={styles.text}
-				opacity={0.8}
+				opacity={0.9}
 			/>
 		);
 	}
